refactor: 태그 필터링 로직 재활용 위해 훅으로 분리
