+++ components/features/place-list/detail/PlaceListDetailWrapper.tsx
@@ -3,6 +3,7 @@
 import { useState } from 'react';
 
 import { QueryBoundary } from '@/components/common/ui/boundary/Queryboundary';
+import { useTagFilter } from '@/hooks/place-list/useTagFilter';
 import { SortType } from '@/types/placeList';
 
 import TagList from '../tag/TagList';
@@ -13,19 +14,7 @@ import SortingDropdownButton from './SortingDropdownButton';
 
 export default function PlaceListDetailWrapper({ listId }: { listId: string }) {
   const [sortBy, setSortBy] = useState<SortType>('created_desc');
-  const [activeTagIds, setActiveTagIds] = useState<Set<string>>(new Set());
-
-  const handleToggleTag = (id: string) => {
-    setActiveTagIds((prev) => {
-      const updated = new Set(prev);
-      if (prev.has(id)) {
-        updated.delete(id);
-      } else {
-        updated.add(id);
-      }
-      return updated;
-    });
-  };
+  const { activeTagIds, handleToggleTag } = useTagFilter();
 
   return (
     <div className='flex flex-col gap-2.5 h-full'>
